requirement.txt
streamlit
requests
beautifulsoup4
python-dotenv                                                                                                                                                                                                                 
transformers
tensorflow

